// src/app/core/enums/enums.ts
export enum UserRoles {
  ROLE_ADMIN = 'ROLE_ADMIN',
  ROLE_PROFESSOR = 'ROLE_PROFESSOR',
  ROLE_STUDENT = 'ROLE_STUDENT'
}

export enum ConfirmOption {
  OK ='OK',
  CANCEL = 'Cancel'
}

export enum Semester {
  SUMMER = 'Summer',
  WINTER = 'Winter'
}

export enum ActiveOptions {
  YES = 'Yes',
  NO = 'No'
}

export enum ModeOptions {
  ADD = 'ADD',
  UPDATE = 'UPDATE'
}